@@ -3,63 +3,72 @@
 source 'https://rubygems.org'
 git_source(:github) { |repo| "https://github.com/#{repo}.git" }
 
-ruby '2.6.3'
+ruby '2.7.2'
 
-gem 'activeadmin'
-gem 'bundler-audit'
-gem 'cloudinary'
-gem 'country_select'
-gem 'devise'
-gem 'devise-security'
-gem 'draper'
+gem 'activeadmin', '~> 2.7.0'
+gem 'cloudinary', '~> 1.13.2'
+gem 'country_select', '~> 4.0.0'
+gem 'devise', '~> 4.7.1'
+gem 'devise-security', '~> 0.14.3'
+gem 'draper', '~> 4.0.1'
+
+# TODO: remove after implementing credentials
 gem 'figaro'
-gem 'haml-rails'
-gem 'i18n'
-gem 'image_processing'
-gem 'lefthook'
-gem 'omniauth-facebook'
+
+gem 'haml-rails', '~> 2.0.1'
+gem 'i18n', '~> 1.8.2'
+gem 'image_processing', '~> 1.10.3'
+gem 'omniauth-facebook', '~> 6.0.0'
 gem 'omniauth-rails_csrf_protection'
-gem 'pagy'
-gem 'pg'
+gem 'pagy', '~> 3.7.5'
+gem 'pg', '~> 1.2.3'
 gem 'puma', '>= 4.3.1'
 gem 'rails', '~> 6.0.1'
-gem 'rspec_junit_formatter'
 gem 'sass-rails', '>= 6'
-gem 'sidekiq'
-gem 'simple_form'
-gem 'site_prism'
+gem 'sidekiq', '~> 6.0.6'
+gem 'simple_form', '~> 5.0.2'
 gem 'turbolinks', '~> 5'
 gem 'webpacker', '~> 4.0'
-gem 'wicked'
-
-gem 'bootsnap', '>= 1.4.2', require: false
+gem 'wicked', '~> 1.3.4'
 
 group :development, :test do
-  gem 'byebug', platforms: %i[mri mingw x64_mingw]
-  gem 'factory_bot_rails'
-  gem 'ffaker'
-  gem 'pry-rails'
-  gem 'rubocop'
-  gem 'shoulda-matchers'
+  gem 'brakeman', '~> 4.10.0'
+  gem 'bullet', '~> 6.1'
+  gem 'bundler-audit', '~> 0.7.0'
+  gem 'bundler-leak', '~> 0.2.0'
+  gem 'database_consistency', '~> 0.8', require: false
+  gem 'factory_bot_rails', '~> 6.1'
+  gem 'fasterer', '~> 0.8.3'
+  gem 'ffaker', '~> 2.14.0'
+  gem 'haml_lint', '~> 0.36', require: false
+  gem 'lefthook', '~> 0.7'
+  gem 'pry-byebug', '~> 3.9'
+  gem 'rails-controller-testing', '~> 1.0'
+  gem 'reek', '~> 6.0'
+  gem 'rspec-rails', '~> 4.0'
+  gem 'rubocop', '~> 0.93', require: false
+  gem 'rubocop-faker', '~> 1.0', require: false
+  gem 'rubocop-performance', '~> 1.9', require: false
+  gem 'rubocop-rails', '~> 2.8', require: false
+  gem 'rubocop-rspec', '~> 1.44', require: false
 end
 
 group :development do
-  gem 'annotate'
-  gem 'brakeman'
-  gem 'letter_opener'
-  gem 'listen', '>= 3.0.5', '< 3.2'
-  gem 'rspec-rails', '~> 4.0.0.beta3'
-  gem 'spring'
-  gem 'spring-watcher-listen', '~> 2.0.0'
-  gem 'web-console', '>= 3.3.0'
+  gem 'i18n-tasks', '~> 0.9'
+  gem 'letter_opener', '~> 1.7'
 end
 
 group :test do
-  gem 'capybara', '>= 2.15'
-  gem 'database_cleaner'
-  gem 'selenium-webdriver'
-  gem 'simplecov'
-  gem 'webdrivers'
+  gem 'capybara', '~> 3.33'
+  gem 'capybara-screenshot', '~> 1.0'
+  gem 'database_cleaner', '~> 1.8'
+  gem 'pundit-matchers', '~> 1.6'
+  gem 'rspec_junit_formatter', '~> 0.4'
+  gem 'rspec-sidekiq', '~> 3.1'
+  gem 'selenium-webdriver', '~> 3.142'
+  gem 'shoulda-matchers', '~> 4.4'
+  gem 'simplecov', '~> 0.19', require: false
+  gem 'simplecov-lcov', '~> 0.8.0', require: false
+  gem 'site_prism', '~> 3.7'
+  gem 'test-prof', '~> 0.12'
 end
-
-gem 'tzinfo-data', platforms: %i[mingw mswin x64_mingw jruby]
