@@ -1,39 +1,8 @@
 # frozen_string_literal: true
 
 require 'open-uri'
-# == Schema Information
-#
-# Table name: customers
-#
-#  id                     :bigint           not null, primary key
-#  avatar                 :string
-#  confirmation_sent_at   :datetime
-#  confirmation_token     :string
-#  confirmed_at           :datetime
-#  current_sign_in_at     :datetime
-#  current_sign_in_ip     :string
-#  email                  :string           default(""), not null
-#  encrypted_password     :string           default(""), not null
-#  last_sign_in_at        :datetime
-#  last_sign_in_ip        :string
-#  provider               :string
-#  remember_created_at    :datetime
-#  reset_password_sent_at :datetime
-#  reset_password_token   :string
-#  sign_in_count          :integer          default(0), not null
-#  uid                    :string
-#  created_at             :datetime         not null
-#  updated_at             :datetime         not null
-#
-# Indexes
-#
-#  index_customers_on_confirmation_token    (confirmation_token) UNIQUE
-#  index_customers_on_email                 (email) UNIQUE
-#  index_customers_on_reset_password_token  (reset_password_token) UNIQUE
-#
-
 class Customer < ApplicationRecord
-  has_many :addresses, as: :addressable
+  has_many :addresses, as: :addressable, dependent: :destroy
   has_many :reviews, dependent: :destroy
   has_many :carts, dependent: :destroy
   has_many :books, through: :carts
@@ -63,7 +32,7 @@ class Customer < ApplicationRecord
   end
 
   def apply_omniauth(auth)
-    update_attributes(
+    update(
       provider: auth.provider,
       uid: auth.uid
     )
