@@ -1,31 +1,8 @@
 # frozen_string_literal: true
 
-# == Schema Information
-#
-# Table name: addresses
-#
-#  id               :bigint           not null, primary key
-#  address          :string           not null
-#  address_type     :integer          not null
-#  addressable_type :string
-#  city             :string           not null
-#  country          :string           not null
-#  first_name       :string           not null
-#  last_name        :string           not null
-#  phone            :string           not null
-#  zipcode          :integer          not null
-#  created_at       :datetime         not null
-#  updated_at       :datetime         not null
-#  addressable_id   :bigint
-#
-# Indexes
-#
-#  index_addresses_on_addressable_type_and_addressable_id  (addressable_type,addressable_id)
-#
-
 class Address < ApplicationRecord
-  REGEX_RULES = { name_on_card: /\A[a-zA-Z]*\s*[a-zA-Z]*\z/, common_name_rule: /\A[a-zA-Z \']*\z/,
-                  address: /\A[a-zA-Z0-9 \-\,\']*\z/, zipcode: /\A[0-9 \-]*\z/, phone: /\A\+[0-9 \-]*\z/ }.freeze
+  REGEX_RULES = { name_on_card: /\A[a-zA-Z]*\s*[a-zA-Z]*\z/, common_name_rule: /\A[a-zA-Z ']*\z/,
+                  address: /\A[a-zA-Z0-9 \-,']*\z/, zipcode: /\A[0-9 \-]*\z/, phone: /\A\+[0-9 \-]*\z/ }.freeze
 
   belongs_to :addressable, polymorphic: true
 
@@ -36,5 +13,5 @@ class Address < ApplicationRecord
   validates :zipcode, format: { with: Address::REGEX_RULES[:zipcode] }, length: { maximum: 10 }
   validates :phone, format: { with: Address::REGEX_RULES[:phone] }, length: { maximum: 15 }
 
-  enum address_type: %i[billing shipping]
+  enum address_type: { billing: 0, shipping: 1 }
 end
