@@ -1,19 +1,7 @@
 # frozen_string_literal: true
 
-# == Schema Information
-#
-# Table name: authors
-#
-#  id         :bigint           not null, primary key
-#  biography  :string
-#  first_name :string           not null
-#  last_name  :string           not null
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
-#
-
 class Author < ApplicationRecord
-  has_many :book_authors
+  has_many :book_authors, dependent: :destroy
   has_many :books, through: :book_authors
 
   validates :first_name, presence: true
