@@ -1,17 +1,7 @@
 # frozen_string_literal: true
 
-# == Schema Information
-#
-# Table name: materials
-#
-#  id         :bigint           not null, primary key
-#  name       :string           not null
-#  created_at :datetime         not null
-#  updated_at :datetime         not null
-#
-
 class Material < ApplicationRecord
-  has_many :books
+  has_many :books, dependent: :destroy
 
   validates :name, presence: true
 end
