@@ -1,15 +1,15 @@
 # frozen_string_literal: true
 
-feature 'Home page' do
+describe 'Home page' do
   let!(:book) { create(:book) }
   let!(:category) { create(:category) }
   let(:home_page) { HomePage.new }
 
-  before :each do
+  before do
     visit root_path
   end
 
-  scenario 'has appropriate content' do
+  it 'has appropriate content' do
     expect(home_page).to have_home_page_link
     expect(home_page).to have_get_started
     expect(home_page).to have_latest_books
@@ -17,20 +17,20 @@ feature 'Home page' do
     expect(home_page).to have_home_page_link
   end
 
-  scenario 'when click Home proceed to Home page' do
+  it 'when click Home proceed to Home page' do
     click_link(I18n.t(:home), match: :first)
 
     expect(page).to have_content(I18n.t(:welcome))
   end
 
-  scenario 'when click Get Started proceed to Catalog page' do
+  it 'when click Get Started proceed to Catalog page' do
     click_link(I18n.t(:shop), match: :first)
     click_link(category.title)
 
     expect(page).to have_content(I18n.t(:catalog))
   end
 
-  scenario 'latest book slider' do
+  it 'latest book slider' do
     expect(home_page).to have_slider
     expect(home_page).to have_latest_books
 
@@ -40,7 +40,7 @@ feature 'Home page' do
     expect(home_page).to have_content(book.title)
   end
 
-  scenario 'bestsellers' do
+  it 'bestsellers' do
     expect(home_page).to have_best_sellers
   end
 end
