@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
-feature 'Book page' do
+describe 'Book page', type: :feature do
   let!(:customer) { create(:customer, :confirmed) }
   let!(:book) { create(:book, :with_author) }
 
-  feature 'allows customer to write a review', js: true do
-    scenario 'successfully creates review' do
+  describe 'allows customer to write a review' do
+    it 'successfully creates review' do
       sign_in(customer)
       visit category_book_path(category_id: book.category_id, id: book.id)
       within('#new_review') do
@@ -17,31 +17,31 @@ feature 'Book page' do
       expect(page).to have_content 'Thanks for Review'
 
       visit category_book_path(category_id: book.category_id, id: book.id)
-      expect(page).to_not have_content 'Review test'
+      expect(page).not_to have_content 'Review test'
       book.reviews.first.approved!
       visit category_book_path book.category_id, book.id
       expect(page).to have_content 'Review test'
     end
   end
 
-  feature 'customer sees' do
+  describe 'customer sees' do
     let!(:review_approved) { create(:review, book: book, customer: customer, status: 'approved') }
     let!(:review_pending) { create(:review, book: book, customer: customer) }
     let!(:review_rejected) { create(:review, book: book, customer: customer, status: 'rejected') }
 
-    scenario 'only approved reviews' do
+    it 'only approved reviews' do
       sign_in(customer)
       visit category_book_path(category_id: book.category_id, id: book.id)
       expect(page).to have_content(review_approved.body)
-      expect(page).to_not have_content(review_pending.body)
-      expect(page).to_not have_content(review_rejected.body)
+      expect(page).not_to have_content(review_pending.body)
+      expect(page).not_to have_content(review_rejected.body)
     end
   end
 
-  feature 'when customer not signed in' do
-    scenario 'unable to leave reviews' do
+  describe 'when customer not signed in' do
+    it 'unable to leave reviews' do
       visit category_book_path(category_id: book.category_id, id: book.id)
-      expect(page).to_not have_content 'Write a Review'
+      expect(page).not_to have_content 'Write a Review'
     end
   end
 end
