@@ -1,32 +1,5 @@
 # frozen_string_literal: true
 
-# == Schema Information
-#
-# Table name: carts
-#
-#  id             :bigint           not null, primary key
-#  number         :string
-#  status         :integer          default("in_progress")
-#  use_billing    :boolean          default(FALSE)
-#  created_at     :datetime         not null
-#  updated_at     :datetime         not null
-#  credit_card_id :bigint
-#  customer_id    :bigint
-#  delivery_id    :bigint
-#
-# Indexes
-#
-#  index_carts_on_credit_card_id  (credit_card_id)
-#  index_carts_on_customer_id     (customer_id)
-#  index_carts_on_delivery_id     (delivery_id)
-#
-# Foreign Keys
-#
-#  fk_rails_...  (credit_card_id => credit_cards.id)
-#  fk_rails_...  (customer_id => customers.id)
-#  fk_rails_...  (delivery_id => deliveries.id)
-#
-
 RSpec.describe Cart do
   describe 'associations' do
     it { is_expected.to have_many(:line_items) }
@@ -38,8 +11,8 @@ RSpec.describe Cart do
 
   describe 'validations' do
     it {
-      is_expected.to define_enum_for(:status).with_values(waiting_for_processing: 0, in_delivery: 1, delivered: 2,
-                                                          canceled: 3)
+      expect(subject).to define_enum_for(:status).with_values(waiting_for_processing: 0, in_delivery: 1, delivered: 2,
+                                                              canceled: 3)
     }
   end
 end
