@@ -16,7 +16,7 @@ require File.expand_path('../config/environment', __dir__)
 abort('The Rails environment is running in production mode!') if Rails.env.production?
 require 'rspec/rails'
 
-Dir[Rails.root.join('spec', 'support', '**', '*.rb')].each(&method(:require))
+Dir[Rails.root.join('spec', 'support', '**', '*.rb')].sort.each { |file| require file }
 
 begin
   ActiveRecord::Migration.maintain_test_schema!
@@ -26,8 +26,6 @@ rescue ActiveRecord::PendingMigrationError => e
 end
 RSpec.configure do |config|
   config.include FactoryBot::Syntax::Methods
-  config.include Devise::Test::ControllerHelpers, type: :controller
-  config.include Warden::Test::Helpers
   config.include CheckoutHelper, type: :feature
   config.include SessionHelper, type: :feature
 
@@ -43,36 +41,7 @@ RSpec.configure do |config|
   config.filter_rails_from_backtrace!
   # arbitrary gems may also be filtered via:
   # config.filter_gems_from_backtrace("gem name")
-  config.after :each do
+  config.after do
     Warden.test_reset!
   end
 end
-Shoulda::Matchers.configure do |config|
-  config.integrate do |with|
-    with.test_framework :rspec
-    with.library :rails
-  end
-end
-
-Capybara.register_driver :site_prism do |app|
-  browser = ENV.fetch('browser', 'chrome').to_sym
-  capabilities = Selenium::WebDriver::Remote::Capabilities.chrome(
-    chromeOptions: { args: %w[headless disable-gpu] }
-  )
-  if ENV['SELENIUM_DRIVER_URL'].present?
-    Capybara::Selenium::Driver.new(
-      app,
-      browser: :remote,
-      url: ENV.fetch('SELENIUM_DRIVER_URL'),
-      desired_capabilities: :chrome
-    )
-  else
-    Capybara::Selenium::Driver.new(app, browser: browser, desired_capabilities: capabilities)
-  end
-end
-
-Capybara.configure do |config|
-  config.default_driver = :site_prism
-end
-
-Capybara.javascript_driver = :site_prism
