@@ -5,15 +5,15 @@ RSpec.configure do |config|
     DatabaseCleaner.clean_with :truncation, except: %w[ar_internal_metadata]
   end
 
-  config.before(:each) do
+  config.before do
     DatabaseCleaner.strategy = :transaction
   end
 
-  config.before(:each) do
+  config.before do
     DatabaseCleaner.start
   end
 
-  config.after(:each) do
+  config.after do
     DatabaseCleaner.clean
   end
 end
